[CHATS-38] Implemented Jenkinsfile
Jenkins pipeline has been defined.

To fix a build issue, configuration has been expanded to also include a properties file configuration. A stub consul configuration was also added.

build-artifacts script was updated to include a no-docker mode when building, this might become useful for future pipeline usages.


Approved-by: Luca Gasparini

+++ build-artifacts.sh
@@ -7,17 +7,18 @@
 function build-all-artifacts() {
   version=$1
   artifacts_folder=$2
-  distro=$3
-  deploy_on=$4
+  no_docker=$3
+  distro=$4
+  deploy_on=$5
 
   declare -a distros=(
-#   "DISTRO | NAME PRE VERSION   | NAME POST VERSION"
-    "centos | carbonio-chats-ce- | -1.el8.x86_64.rpm"
-    "ubuntu | carbonio-chats-ce_ | -1_amd64.deb     "
+    #   "DISTRO  | NAME PRE VERSION   | NAME POST VERSION"
+    "rocky-8 | carbonio-chats-ce- | -1.el8.x86_64.rpm"
+    "ubuntu  | carbonio-chats-ce_ | -1_amd64.deb     "
   )
   distro_found=false
   for distros_index in "${!distros[@]}"; do
-    IFS=' | ' read -r -a distros_item <<< "${distros[distros_index]}"
+    IFS=' | ' read -r -a distros_item <<<"${distros[distros_index]}"
     if [[ "$distro" == "${distros_item[0]}" || "$distro" == "" ]]; then
       distro_found=true
       print-banner "Building ${distros_item[0]} package"
@@ -64,44 +65,65 @@ EOF
 }
 
 function build-ubuntu-artifact() {
+  if [ "$no_docker" = true ]; then
+    mkdir /tmp/chats
+    cp * /tmp/chats
+    cd /tmp/chats || exit
+    pacur build ubuntu
+  else
     docker run \
       --rm --entrypoint "" \
       -v "$(pwd)":/tmp/chats \
       -e VERSION="$1" \
       registry.dev.zextras.com/jenkins/pacur/ubuntu-18.04:v1 /bin/bash -c 'cd /tmp/chats && pacur build ubuntu'
+  fi
 }
 
-function build-centos-artifact() {
-  docker run \
-    --rm --entrypoint "" \
-    -v "$(pwd)":/tmp/chats \
-    -e VERSION="$1" \
-    registry.dev.zextras.com/jenkins/pacur/centos-8:v1 /bin/bash -c 'cd /tmp/chats && pacur build centos'
+function build-rocky-8-artifact() {
+  if [ "$no_docker" = true ]; then
+    mkdir /tmp/chats
+    cp * /tmp/chats
+    cd /tmp/chats || exit
+    pacur build rocky-8
+  else
+    docker run \
+      --rm --entrypoint "" \
+      -v "$(pwd)":/tmp/chats \
+      -e VERSION="$1" \
+      registry.dev.zextras.com/jenkins/pacur/rocky-8:v1 /bin/bash -c 'cd /tmp/chats && pacur build rocky-8'
+  fi
 }
 
 function print-banner() {
   string_to_print=$1
   banner_string=""
   if [ ${#string_to_print} -lt 60 ]; then
-    start_spaces=$((60-${#string_to_print}))
-    start_spaces=$((start_spaces/2))
+    start_spaces=$((60 - ${#string_to_print}))
+    start_spaces=$((start_spaces / 2))
     index=0
-    while [ $index -lt $start_spaces ]; do banner_string="$banner_string "; index=$((index+1)); done;
+    while [ $index -lt $start_spaces ]; do
+      banner_string="$banner_string "
+      index=$((index + 1))
+    done
     banner_string="$banner_string$string_to_print"
-    index=$((index+${#string_to_print}))
-    while [ $index -lt 60 ]; do banner_string="$banner_string "; index=$((index+1)); done;
+    index=$((index + ${#string_to_print}))
+    while [ $index -lt 60 ]; do
+      banner_string="$banner_string "
+      index=$((index + 1))
+    done
   else
     banner_string="$string_to_print"
   fi
   index=0
   border_string=""
-  while [ $index -lt 72 ]; do border_string="$border_string*"; index=$((index+1)); done;
+  while [ $index -lt 72 ]; do
+    border_string="$border_string*"
+    index=$((index + 1))
+  done
   echo ""
   echo "$border_string"
   echo "****  $banner_string  ****" | tr a-z A-Z
   echo "$border_string"
 }
 
-build-all-artifacts "$1" "$2" "$3" "$4"
-
-
+build-all-artifacts "$1" "$2" "$3" "$4" "$5"
